Bmp280 floats (#52)
* [IMPROVEMENT] changed int to float for bmp280
Description:
- removed old files for bmp280
- now bmp280 module returns floats instead of ints
- increased buffer len to 10

+++ bike_computer_v3/source/drivers/BMP280/include/bmp280.hpp
@@ -5,7 +5,7 @@
 
 namespace bmp280{
     void init();
-    std::tuple<int32_t ,int32_t> get_temp_press();
+    std::tuple<float ,float> get_temp_press();
     /**
      * @brief calculates height above the sea with msl pressure, current pressure and temp
      *
